+++ dash-abci/src/execution/mod.rs
@@ -1,3 +1,4 @@
 mod distribution_storage_fees;
 mod process_block_fees;
 mod block_fee_distribution;
+pub mod fee_distribution;
