+++ drive/src/common/mod.rs
@@ -15,6 +15,7 @@ use grovedb::TransactionArg;
 
 use crate::contract::Contract;
 use crate::drive::Drive;
+use crate::drive::flags::StorageFlags;
 use crate::error::structure::StructureError;
 use crate::error::Error;
 
@@ -28,7 +29,7 @@ pub fn setup_contract(
     let contract =
         Contract::from_cbor(&contract_cbor, contract_id).expect("contract should be deserialized");
     drive
-        .apply_contract_cbor(contract_cbor, contract_id, 0f64, true, transaction)
+        .apply_contract_cbor(contract_cbor, contract_id, 0f64, true, StorageFlags::default(), transaction)
         .expect("contract should be applied");
     contract
 }
@@ -42,7 +43,7 @@ pub fn setup_contract_from_hex(
     let contract =
         Contract::from_cbor(&contract_cbor, None).expect("contract should be deserialized");
     drive
-        .apply_contract_cbor(contract_cbor, None, 0f64, true, transaction)
+        .apply_contract_cbor(contract_cbor, None, 0f64, true,StorageFlags::default(),  transaction)
         .expect("contract should be applied");
     contract
 }
