+++ drive/src/drive/document/update.rs
@@ -32,6 +32,7 @@ impl Drive {
         owner_id: Option<&[u8]>,
         block_time: f64,
         apply: bool,
+        storage_flags: StorageFlags,
         transaction: TransactionArg,
     ) -> Result<(i64, u64), Error> {
         let contract = Contract::from_cbor(contract_cbor, None)?;
@@ -46,6 +47,7 @@ impl Drive {
             owner_id,
             block_time,
             apply,
+            storage_flags,
             transaction,
         )
     }
@@ -58,6 +60,7 @@ impl Drive {
         owner_id: Option<&[u8]>,
         block_time: f64,
         apply: bool,
+        storage_flags: StorageFlags,
         transaction: TransactionArg,
     ) -> Result<(i64, u64), Error> {
         let document = Document::from_cbor(serialized_document, None, owner_id)?;
@@ -70,6 +73,7 @@ impl Drive {
             owner_id,
             block_time,
             apply,
+            storage_flags,
             transaction,
         )
     }
@@ -83,19 +87,12 @@ impl Drive {
         owner_id: Option<&[u8]>,
         block_time: f64,
         apply: bool,
+        storage_flags: StorageFlags,
         transaction: TransactionArg,
     ) -> Result<(i64, u64), Error> {
         let mut drive_operations: Vec<DriveOperation> = vec![];
 
         let document_type = contract.document_type_for_name(document_type_name)?;
-
-        let block_execution_context = self.block_execution_context.borrow();
-        let epoch = match block_execution_context.deref() {
-            Some(block_execution_context) => block_execution_context.epoch_info.current_epoch_index,
-            None => 0,
-        };
-
-        let storage_flags = StorageFlags { epoch };
 
         let document_info = if apply {
             DocumentAndSerialization((document, serialized_document, &storage_flags))
@@ -504,6 +501,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to apply contract successfully");
@@ -521,6 +519,7 @@ mod tests {
                 true,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("should create alice profile");
@@ -537,6 +536,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("should update alice profile");
@@ -556,7 +556,7 @@ mod tests {
         let contract = Contract::from_cbor(contract_cbor.as_slice(), None)
             .expect("expected to create contract");
         drive
-            .apply_contract_cbor(contract_cbor.clone(), None, 0f64, true, None)
+            .apply_contract_cbor(contract_cbor.clone(), None, 0f64, true, StorageFlags::default(), None)
             .expect("expected to apply contract successfully");
 
         // Create Alice profile
@@ -612,6 +612,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 None,
             )
             .expect("should update alice profile");
@@ -644,6 +645,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to apply contract successfully");
@@ -715,6 +717,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("should update alice profile");
@@ -753,6 +756,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to apply contract successfully");
@@ -852,6 +856,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("should update alice profile");
@@ -902,7 +907,7 @@ mod tests {
         let contract = value_to_cbor(contract, Some(defaults::PROTOCOL_VERSION));
 
         drive
-            .apply_contract_cbor(contract.clone(), None, 0f64, true, None)
+            .apply_contract_cbor(contract.clone(), None, 0f64, true, StorageFlags::default(), None)
             .expect("should create a contract");
 
         // Create document
@@ -931,6 +936,7 @@ mod tests {
                 true,
                 0f64,
                 true,
+                StorageFlags::default(),
                 None,
             )
             .expect("should add document");
@@ -960,6 +966,7 @@ mod tests {
                 None,
                 0f64,
                 true,
+                StorageFlags::default(),
                 None,
             )
             .expect("should update document");
@@ -1015,6 +1022,7 @@ mod tests {
                 false,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to insert a document successfully");
@@ -1027,6 +1035,7 @@ mod tests {
                 Some(&random_owner_id),
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect_err("expected not to be able to update a non mutable document");
@@ -1040,6 +1049,7 @@ mod tests {
                 true,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect_err("expected not to be able to override a non mutable document");
@@ -1083,6 +1093,7 @@ mod tests {
                 false,
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to insert a document successfully");
@@ -1095,6 +1106,7 @@ mod tests {
                 Some(&random_owner_id),
                 0f64,
                 true,
+                StorageFlags::default(),
                 Some(&db_transaction),
             )
             .expect("expected to update a document with history successfully");
@@ -1159,6 +1171,7 @@ mod tests {
         let config = DriveConfig {
             batching_enabled: using_batches,
             has_raw_enabled: using_has_raw,
+            default_genesis_time: 0,
             encoding: DriveEncoding::DriveCbor,
         };
         let tmp_dir = TempDir::new().unwrap();
